--- no_code_variable_options.go
@@ -1,18 +1,0 @@
-package tfe
-
-// NoCodeVariableOption represents a registry no-code module variable
-// option.
-type NoCodeVariableOption struct {
-	// Type is a public field utilized by JSON:API to
-	// set the resource type via the field tag.
-	Type string `jsonapi:"primary,variable-options"`
-
-	// Required: The variable name
-	VariableName string `jsonapi:"attr,variable-name"`
-
-	// Required: The variable type
-	VariableType string `jsonapi:"attr,variable-type"`
-
-	// Optional: The options for the variable
-	Options []string `jsonapi:"attr,options"`
-}
