+++ registry_no_code_module_integration_test.go
